Convert back to old method of picking institution
Prototype autocomplete was present. However, it did nothing to address
that the character set usage for list of institutions is absolutely
crazy. So people may or may not be using the correct accented variation
of their institution name, and we may or may not correctly match. Really
needs to use an embedded version of Solr so that the index and search
values follow the same manipulation rules.